Cleanup TryFetchAndLogKextHealthData

+++ ProjFS.Mac/PrjFSKext/public/PrjFSHealthData.h
@@ -2,5 +2,5 @@
 
 struct PrjFSHealthData
 {
-    uint64_t cacheSize;
+    uint32_t cacheSize;
 };
